refactor(tunnel): extract sans-IO core into tunnel-proto crate
Move the synchronous Client/Gateway state machine and its supporting modules (client, gateway, messages, dns, filter_engine, routing, ...) out of `tunnel` into a new `tunnel-proto` crate. `tunnel` keeps the IO event loop (`Tunnel`, `io`, `sockets`) and re-exports `tunnel-proto`, so downstream crates are unchanged.

The state-machine methods the event loop drives are now plain `pub` instead of `visibility::make(pub)` behind test-util; the `visibility` dep drops out. This isolates the coverable sans-IO surface so coverage can target `tunnel-proto` directly.
