+++ rust/Cargo.toml
@@ -30,6 +30,7 @@ members = [
     "libs/connlib/socket-factory",
     "libs/connlib/tun",
     "libs/connlib/tunnel",
+    "libs/connlib/tunnel-proto",
     "libs/eventloop-budget",
     "libs/flow-log-spool",
     "libs/flow-log-upload",
@@ -240,6 +241,7 @@ tracing-subscriber = { version = "0.3.23", features = ["parking_lot"] }
 tun = { path = "libs/connlib/tun" }
 tunnel = { path = "libs/connlib/tunnel" }
 tunnel-bypass-resolver = { path = "libs/tunnel-bypass-resolver" }
+tunnel-proto = { path = "libs/connlib/tunnel-proto" }
 uniffi = "0.31.1"
 url = "2.5.2"
 uuid = "1.23.2"
