+++ rust/libs/connlib/tunnel-proto/Cargo.toml
@@ -1,0 +1,63 @@
+[package]
+name = "tunnel-proto"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+
+[features]
+proptest = ["dep:proptest"]
+divan = ["dep:divan"]
+# Exposes test-only internals and enables the `proptest` strategies used by
+# downstream test harnesses. Not meant for production use.
+test-util = ["proptest"]
+
+[dependencies]
+anyhow = { workspace = true }
+base64 = { workspace = true, features = ["std"] }
+bimap = { workspace = true }
+boringtun = { workspace = true }
+bytes = { workspace = true, features = ["std"] }
+chrono = { workspace = true }
+connlib-model = { workspace = true }
+derive_more = { workspace = true, features = ["debug", "from", "display"] }
+divan = { workspace = true, optional = true }
+dns-over-tcp = { workspace = true }
+dns-types = { workspace = true }
+flow-tracker = { workspace = true }
+glob = { workspace = true }
+ip-packet = { workspace = true }
+ip_network = { workspace = true }
+ip_network_table = { workspace = true }
+itertools = { workspace = true, features = ["use_std"] }
+l3-udp-dns-client = { workspace = true }
+logging = { workspace = true }
+lru = { workspace = true }
+opentelemetry = { workspace = true, features = ["metrics"] }
+otel-instruments = { workspace = true }
+proptest = { workspace = true, optional = true }
+rand = { workspace = true }
+rangemap = { workspace = true }
+ringbuffer = { workspace = true }
+secrecy = { workspace = true, features = ["serde"] }
+serde = { workspace = true, features = ["derive", "std"] }
+serde_json = { workspace = true }
+serde_with = { workspace = true }
+smallvec = { workspace = true, features = ["const_generics"] }
+snownet = { workspace = true }
+telemetry = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true, features = ["attributes"] }
+
+[dev-dependencies]
+ip-packet = { workspace = true, features = ["proptest"] }
+rand = { workspace = true }
+test-case = { workspace = true }
+test-strategy = { workspace = true }
+
+[[bench]]
+name = "divan"
+harness = false
+required-features = ["divan"]
+
+[lints]
+workspace = true
