+++ rust/libs/connlib/tunnel-proto/benches/divan.rs
@@ -1,0 +1,5 @@
+extern crate tunnel_proto; // Ensure benchmarks aren't optimised out.
+
+fn main() {
+    divan::main()
+}
