+++ rust/libs/connlib/tunnel-proto/src/client/dns_config.rs
@@ -55,14 +55,14 @@ impl DnsMapping {
     // For such small numbers, linear search is usually more efficient.
     // Most importantly, it is much easier for us to retain the ordering of the DNS servers if we don't use a map.
 
-    #[cfg(all(test, feature = "proptest"))]
-    pub(crate) fn sentinel_by_upstream(&self, upstream: &dns::Upstream) -> Option<IpAddr> {
+    #[cfg(feature = "test-util")]
+    pub fn sentinel_by_upstream(&self, upstream: &dns::Upstream) -> Option<IpAddr> {
         self.inner
             .iter()
             .find_map(|(sentinel, candidate)| (candidate == upstream).then_some(*sentinel))
     }
 
-    pub(crate) fn upstream_by_sentinel(&self, sentinel: IpAddr) -> Option<dns::Upstream> {
+    pub fn upstream_by_sentinel(&self, sentinel: IpAddr) -> Option<dns::Upstream> {
         self.inner
             .iter()
             .find_map(|(candidate, upstream)| (candidate == &sentinel).then_some(upstream.clone()))
