+++ rust/libs/connlib/tunnel-proto/src/client/dns_resource_nat.rs
