+++ rust/libs/connlib/tunnel-proto/src/client/pending_authorizations.rs
@@ -265,7 +265,7 @@ fn is_trigger_allowed(trigger: &Trigger, filter: &FilterEngine) -> bool {
         return true;
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-util"))]
     if crate::malicious_behaviour::ignore_resource_filter() {
         tracing::debug!("Malicious client: ignoring resource filter");
         return true;
