+++ rust/libs/connlib/tunnel-proto/src/gateway.rs
@@ -73,7 +73,7 @@ impl DnsResourceNatEntry {
 }
 
 impl GatewayState {
-    pub(crate) fn new(seed: [u8; 32], now: Instant, unix_ts: Duration) -> Self {
+    pub fn new(seed: [u8; 32], now: Instant, unix_ts: Duration) -> Self {
         Self {
             peers: Default::default(),
             node: Node::new(seed, now, unix_ts),
@@ -91,12 +91,12 @@ impl GatewayState {
         self.flow_tracker.set_enabled(enabled);
     }
 
-    #[cfg(all(test, feature = "proptest"))]
-    pub(crate) fn tunnel_ip_config(&self) -> Option<IpConfig> {
+    #[cfg(feature = "test-util")]
+    pub fn tunnel_ip_config(&self) -> Option<IpConfig> {
         self.tun_ip_config
     }
 
-    pub(crate) fn public_key(&self) -> PublicKey {
+    pub fn public_key(&self) -> PublicKey {
         self.node.public_key()
     }
 
@@ -109,7 +109,7 @@ impl GatewayState {
     }
 
     /// Handles packets received on the TUN device.
-    pub(crate) fn handle_tun_input(
+    pub fn handle_tun_input(
         &mut self,
         packet: IpPacket,
         now: Instant,
@@ -151,7 +151,7 @@ impl GatewayState {
     /// Some of them will however be handled internally, for example, TURN control packets exchanged with relays.
     ///
     /// In case this function returns `None`, you should call [`GatewayState::handle_timeout`] next to fully advance the internal state.
-    pub(crate) fn handle_network_input(
+    pub fn handle_network_input(
         &mut self,
         local: SocketAddr,
         from: SocketAddr,
@@ -535,13 +535,13 @@ impl GatewayState {
         }
     }
 
-    pub(crate) fn poll_transmit(&mut self) -> Option<snownet::Transmit> {
+    pub fn poll_transmit(&mut self) -> Option<snownet::Transmit> {
         self.buffered_transmits
             .poll_transmit()
             .or_else(|| self.node.poll_transmit())
     }
 
-    pub(crate) fn poll_event(&mut self) -> Option<GatewayEvent> {
+    pub fn poll_event(&mut self) -> Option<GatewayEvent> {
         if let Some(ev) = self.buffered_events.pop_front() {
             return Some(ev);
         }
