+++ rust/libs/connlib/tunnel-proto/src/utils.rs
