--- rust/libs/connlib/tunnel/benches/divan.rs
@@ -1,5 +1,0 @@
-extern crate tunnel; // Ensure benchmarks aren't optimised out.
-
-fn main() {
-    divan::main()
-}
